apps/pyproject.toml
[project]
name = "ticketforge-ingest"
version = "0.1.0"
description = "GitHub issue ingestion for TicketForge"
requires-python = ">=3.10"

dependencies = [
    "PyGithub",
    "python-dotenv",
    "pandas"
]
